% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPolyTools.R
\name{calcRstats}
\alias{calcRstats}
\title{Calculate statistics of R per marker}
\usage{
calcRstats(datpoly, out)
}
\arguments{
\item{datpoly}{a data frame in long format (input format for fitPoly)
with at least columns MarkerName, SampleName, R}

\item{out}{name of output file; if NA no file is written. File contains
the same data as the return value}
}
\value{
a data frame with columns MarkerName, mean, min, q50, q95, max
which are all statistics of the R values per marker
}
\description{
Calculate the min, max, mean and 50% and 95% quantiles of R
for each marker
}
\details{
The data.frame returned by this function is used as input for
function selMarkers_byR
}
