% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPolyTools.R
\name{checkFilename}
\alias{checkFilename}
\title{Check if a file can be created}
\usage{
checkFilename(filename, overwrite=TRUE)
}
\arguments{
\item{filename}{a file name with or without path. If the name contains a
path the entire path must already exist, else the file cannot be created
and the result is FALSE.}

\item{overwrite}{if TRUE an attempt is made to delete any pre-existing file
of that name; the function returns FALSE if the file is locked or the user
has no rights to delete the file. If FALSE the function returns FALSE
if any file of that name already exists, and no attempt is made to remove it}
}
\value{
TRUE if a new file can be created, else FALSE. If TRUE, no file of
that name exists (any more)
}
\description{
Checks if a file with that name can be created.
If successful, any pre-existing file does not exist any more.
}
