% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPoly.r
\docType{package}
\name{fitPoly}
\alias{fitPoly-package}
\alias{fitPoly}
\title{fitPoly: a package for assigning dosage scores based on SNP array data}
\description{
fitPoly (an evolved version of package fitTetra) fits mixture models to
the distribution of intensity ratios Y/(X+Y) (where X and Y are the
intensities of the signals produced by the A and B alleles of bi-allelic
markers) and uses these to assign genotypes (dosages).
The main differences compared with fitTetra are that it can handle
any ploidy level, and multiple populations that can be either F1 populations
(and their parents) or panels of accessions. There are also improvements
in accuracy, speed and the possibility to use prior dosage information.
}
