% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPolyTools.R
\name{removeRedundant}
\alias{removeRedundant}
\title{Remove redundant single-probe markers}
\usage{
removeRedundant(compstat, combscores, compfile, combscorefile)
}
\arguments{
\item{compstat}{a data.frame as returned by compareProbes in the
compstat item of the return value and in the compfile}

\item{combscores}{a data.frame as returned by compareProbes in the}

\item{compfile}{a filename to which the compstat part of the results is
written; if NA no file is written}

\item{combscorefile}{a filename to which the compscores part of the results is
written; if NA no file is written}
}
\value{
A list with two components, compstat and combscores. These are identical
to the parameters compstat and combscores, with the redundant single-probe
markers removed.For their contents see function compareProbes.
}
\description{
If for a SNP both probes produced very similar results, function
compareProbes generated an extra marker combining the results from both
probes. In that case the original single-probe markers are redundant and this
function removes those data from the compstat and combscores data frames.
}
