% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPoly.r
\name{saveMarkerModels}
\alias{saveMarkerModels}
\title{DEPRECATED: Function to fit mixture models for series of markers and save the
results to files}
\usage{
saveMarkerModels(..., p.threshold = 0.99)
}
\arguments{
\item{...}{All parameters allowed in the function `fitMarkers()`. For a full description
see the help of that function.}

\item{p.threshold}{The minimum P-value required to assign a genotype (dosage)
to a sample; default 0.99 (very stringent). If the P-value for all possible genotypes is less
than p.threshold the sample is assigned genotype NA. Set to 1 to disable
this filter.}
}
\value{
See `fitMarkers()` documentation for a full description.
}
\description{
This is the old name of the function `fitMarkers()`, the main function
of `fitPoly`. It is kept only for backwards-compatibility. The only difference between
the two is the default parameter of the genotyping probability threshold `p.threshold`, 
it is 0.99 in `saveMarkerModels()` (as was originally set) and it is 0.9 in the current
`fitMarkers()`.
}
