% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPolyTools.R
\docType{data}
\name{scores}
\alias{scores}
\title{A data set with dosage scores generated by fitPoly}
\usage{
data(scores)
}
\description{
A data frame as saved by function fitMarkers in package
fitPoly
}
\details{
scores contains the following columns:
\itemize{
 \item marker: numeric, the sequential number of the markers in the
 fitMarkers input file
 \item MarkerName: factor, names of the markers (SNPs)
 \item SampleName: factor names of the individuals / DNA samples
 \item ratio: numeric, the ratios of the Y signal to the total signal
 intensity
 \item P0 ... P4: numeric, the probabilities that the observed signal ratio
 reflects a true Y allele dosage of 0 ... 4, according to the selected
 mixture model
 \item maxgeno: numeric, the dosage with the maximum probability
 \item maxP: numeric, the maximum probability, i.e. the probability of maxgeno
 \item geno: numeric, equal to maxgeno if maxP was above the specified threshold
 (p.threshold=0.99), else NA
}
}
