% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPolyTools.R
\docType{data}
\name{XYdat}
\alias{XYdat}
\title{A data set containing SNP array data}
\usage{
data(XYdat)
}
\description{
A data frame containing SNP array signal intensity data
}
\details{
XYdat contains the following columns:
\itemize{
 \item MarkerName: factor, names of the markers (SNPs)
 \item SampleName: factor names of the individuals / DNA samples
 \item X: numeric, the signal intensities of the X allele
 \item Y: numeric, the signal intensities of the Y allele
 \item R: numeric, the total signal intensities; R = X + Y
 \item ratio: numeric, the ratios of the Y signal to the total signal
 intensity; ratio = Y / R
}
}
